% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_plotting.R
\name{cluster_identities}
\alias{cluster_identities}
\title{Generate N number of piecharts, where N is the number of factor levels of the seurat clusters or a unique metadata feature.}
\usage{
cluster_identities(
  object,
  meta_feature,
  plot_by = "clusters",
  facet = TRUE,
  include_labels = TRUE,
  label_by = "counts"
)
}
\arguments{
\item{object}{An existing Seurat object.}

\item{meta_feature}{The metadata feature that is highlighted when plotting clusters.}

\item{plot_by}{Character string of either 'clusters' or 'meta_feature', denoting what to split by.}

\item{facet}{Boolean value to denote if the plot should have facet_wrap applied.}

\item{include_labels}{Boolean value to denote if the plot should have labels for each slice.}

\item{label_by}{Character string, either "counts" or "percent"}
}
\description{
Generate N number of piecharts, where N is the number of factor levels of the seurat clusters or a unique metadata feature.
}
\examples{
# Create a pie chart with labels and apply facet_wrap with custom parameters after
plot <- cluster_identities(object = seurat_object, meta_feature = "collection.date", plot_by = 'clusters', facet = FALSE, include_labels = TRUE)
plot + facet_wrap(~ combined)
}
