% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{ensemble_to_id}
\alias{ensemble_to_id}
\title{Convert a given list of ensemble transcript id's into \code{ensembl_transcript_id}'s, \code{external_gene_name}'s, \code{ensembl_gene_id}'s, and the original transcript.id}
\usage{
ensemble_to_id(sig_results)
}
\arguments{
\item{sig_results}{A data frame or list containing ensemble transcript id's}
}
\value{
Converts the original data frame or list into a data frame with columns corresponding to \code{ensembl_transcript_id}'s, \code{external_gene_name}'s, \code{ensembl_gene_id}'s, and the original transcript.id
}
\description{
Convert a given list of ensemble transcript id's into \code{ensembl_transcript_id}'s, \code{external_gene_name}'s, \code{ensembl_gene_id}'s, and the original transcript.id
}
\examples{
# Convert a data frame of ensemble transcript id's
ensemble_to_id(wald_sig_results)
}
