% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{f05}
\alias{f05}
\title{Retrieve all target id's for a given data frame with a FDR <.05}
\usage{
f05(sleuth_res)
}
\arguments{
\item{sleuth_res}{A data frame generated following running sleuth_results}
}
\value{
A list of target id's.
}
\description{
Retrieve all target id's for a given data frame with a FDR <.05
}
\examples{
# Given a Sleuth object, run a Likewise Ratio test on all possible models derived from the original formula used to run sleuth_prep
f05(wald_test_results)
}
