% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_statistical_tests.R
\name{filtered_scaled_transcript_counts}
\alias{filtered_scaled_transcript_counts}
\title{Return the scaled transcript counts for a set of genes.}
\usage{
filtered_scaled_transcript_counts(sleuth_obj, genes)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit().}

\item{genes}{A vector containing gene names as strings that should be returned.}
}
\value{
Filtered scaled transcript counts for a specific set of genes.
}
\description{
Return the scaled transcript counts for a set of genes.
}
\examples{
# Given a Sleuth object, retrieve the filtered scaled transcript counts for all the sample for a given gene list.
filtered_scaled_transcript_counts(so_nominal, "ABH1")
}
