% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{isoform_boxplot}
\alias{isoform_boxplot}
\title{Generate a box plot from kruskal-wallis results with user-selected title.}
\usage{
isoform_boxplot(kruskal_result, graph_name = "Sample Plot Name")
}
\arguments{
\item{kruskal_result}{A sleuth object.}

\item{graph_name}{Name of the graph. Default is "Sample Plot Name".}
}
\description{
Generate a box plot from kruskal-wallis results with user-selected title.
}
\examples{
# Create a box plot from a kruskal-wallis result retrived using sleuth_kruskal_wallis(). 
isoform_boxplot(kw_result, graph_name = "All Isoforms")
}
