% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_cell_plotting.R
\name{plotFromReducedDim}
\alias{plotFromReducedDim}
\title{Generate a multidimensional PCA plot with independent loading plots.}
\usage{
plotFromReducedDim(
  seu_object = NULL,
  ggPlotList = NULL,
  nloadings = 5,
  loadingScaleFactor = 150
)
}
\arguments{
\item{ggPlotList}{The output graph generated by plotReducedDim.}

\item{nloadings}{A numeric value for the number of loadings to showcase on single loading plot.}

\item{loadingScaleFactor}{A numeric value to scale the loading vectors by.}

\item{object}{An existing Seurat object.}
}
\description{
Generate a multidimensional PCA plot with independent loading plots.
}
\examples{
# Create a multidimensional PCA plot with independent loading plots
prd_obj <- plotReducedDim(sce_object, dimred = "PCA", colour_by = "seurat_clusters", ncomponents = 3)
plotFromReducedDim(seu_object, prd_obj, nloadings = 5, loadingScaleFactor = 100)
}
