% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_statistical_tests.R
\name{sleuth_kruskal_wallis}
\alias{sleuth_kruskal_wallis}
\title{Run a Kruskal-Wallis One-Way ANOVA on non-parametric data at the isoform level
for a selected gene. The quantile of data the test is run upon is calculated
from the the est_counts of all the data and the resulting value is used to
filter based upon the mean of the est_counts for each individual transcript.
For example, for a quantile probability of 0.25, the function will determine
the quantile probability for 0.25 and assume this value to be 50. The
function will then determine the mean est_counts for each transcript of the
gene and if the mean is greater than the quantile calculated value of 50, it
will be included when running the test.}
\usage{
sleuth_kruskal_wallis(sleuth_obj, gene, grouping, iqf = 0, threshold = 0.05)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit().}

\item{gene}{The name of a gene as a string to asses the isoform expression level between samples.}

\item{grouping}{A column name in the sample to covariate data frame provided when running sleuth_prep / sleuth_interpret.}

\item{iqf}{A numeric value between 0 and 1 (inclusive) to denote the quantile of data to run the test upon.}

\item{threshold}{A numeric value for the p-value cutoff when running the resulting pairwise comparison.}
}
\value{
Results of a kruskal-wallis test at the isoform level.
}
\description{
Run a Kruskal-Wallis One-Way ANOVA on non-parametric data at the isoform level
for a selected gene. The quantile of data the test is run upon is calculated
from the the est_counts of all the data and the resulting value is used to
filter based upon the mean of the est_counts for each individual transcript.
For example, for a quantile probability of 0.25, the function will determine
the quantile probability for 0.25 and assume this value to be 50. The
function will then determine the mean est_counts for each transcript of the
gene and if the mean is greater than the quantile calculated value of 50, it
will be included when running the test.
}
\examples{
# Given a Sleuth object, run the Kruskal-Wallis test of the est_counts factor between isoform groups for all samples.
sleuth_kruskal_wallis(so_nominal, "ABH1", "sex", iqf = 0.25)
}
