% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_model_rss}
\alias{sleuth_model_rss}
\title{Calculate and return the sum of the RSS values available for the model, model factors, and the lrt factors (if the LRT was ran).}
\usage{
sleuth_model_rss(
  sleuth_obj,
  include_wald = FALSE,
  include_lrt = FALSE,
  return_df = FALSE
)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit().}

\item{include_wald}{Boolean value to indicate if the wald test results should be returned. Default is FALSE.}

\item{include_lrt}{Boolean value to indicate if the wald test results should be returned. Default is FALSE.}

\item{return_df}{Boolean value to indicate if the result should be returned as a dataframe or a list. If FALSE, then return just the values.}
}
\value{
Dataframe or characters with all the RSS values for a given sleuth_object.
}
\description{
Calculate and return the sum of the RSS values available for the model, model factors, and the lrt factors (if the LRT was ran).
}
\examples{
# Given a Sleuth object, retrieve the RSS values.
sleuth_model_rss(so_nominal)
}
