% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_object_result}
\alias{sleuth_object_result}
\title{Retrieve all results from a given Sleuth object following statistical testing}
\usage{
sleuth_object_result(
  sleuth_obj,
  all_data = TRUE,
  sig_data = TRUE,
  q_max = 0.05,
  test = "wt"
)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit()}

\item{all_data}{A Boolean value that indicates whether all results of the Wald test should be assigned to a variable. Defaults to TRUE.}

\item{sig_data}{A Boolean value that indicates whether all the significant results of the Wald test coupled with fdr_cutoff should be assigned to a variable. Defaults to TRUE.}

\item{q_max}{The FDR cutoff value that is passed to the fdr_cutoff function embedded within sleuth_object_result. Defaults to .05.}

\item{test}{The sleuth test to be analyzed using sleuth_results, can be either "wt" or "lrt". Defaults to "wt".}
}
\value{
Firstly, newly generated variable containing resulting data from Wald or Likewise Ratio tests performed. Secondly, data frame containing data points with an FDR <.05
}
\description{
Retrieve all results from a given Sleuth object following statistical testing
}
\examples{
# Given a Sleuth object and Boolean values of tests to run, perform the requested tests on all possible models for the given Sleuth object
sleuth_object_result(so, q_max = .01, test = "lrt")
}
