% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumshoe.R
\name{sleuth_prep_transcript_dosage}
\alias{sleuth_prep_transcript_dosage}
\title{Add a transcript's scaled est_count value as a numeric dosage factor on top of an existing Sleuth object's model}
\usage{
sleuth_prep_transcript_dosage(sleuth_obj, metadata, transcript_id)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep()}

\item{metadata}{The metadata that was used to populate sleuth_obj}

\item{transcript_id}{The name of the transcript to model quantitatively}
}
\value{
A new Sleuth object with the new regression model
}
\description{
Can be used to correct for transcript noise/artefacts in a model when it is proportional to a biomarker transcript.
}
\examples{
# Read in metadata with "sample", "path" and other columns that can be used as regression model factors like "treatment"
meta <- read.table("tab_delimited_file_with sample_and_path_columns_etc.txt", header=TRUE)
meta$path <- as.character(meta$path)
so <- sleuth_prep(meta, ~treatment)
so <- sleuth_fit(so)
pca_plot(so)
Plot shows one sample way off on the X-axis, with biggest loading vector in the X-shift direction being NM_001402 (EEF1A1, a biomaker for cell proliferation)
The following call builds a new differential expression model (~treatment + NM_001402) that accounts for this artefact.
so_redux <- sleuth_prep_transcript_dosage(so, meta, "NM_001402")

Compare the model fits (lower number os better)
sleuth_model_rss(so)
sleuth_model_rss(so_redux)
}
\keyword{correction}
\keyword{dosage}
\keyword{error}
\keyword{model,}
