% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumshoe.R
\name{sleuth_reliable_target_ids}
\alias{sleuth_reliable_target_ids}
\title{List all the Sleuth object \code{target_id}s that have at least \code{min_count_per_sample} magnitude in the scaled \code{est_count}.}
\usage{
sleuth_reliable_target_ids(sleuth_obj, min_count_per_sample = 1)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep().}

\item{min_count_per_sample}{threshold that all scaled \code{est_count}s for a transcript should have to be included im the results.}
}
\value{
a list of target_ids meeting the given criterion.
}
\description{
Can be used to provide a non-default target_ids paramater value in other methods of this package, effectively reducing the
magnitude of multiple testing correction in statistical tests to restricting the reported tests to those transcripts with replicable
expression levels (not potentially zero-inflated).
}
\examples{
# Check the replicate dot plots for a factor combination (typically the control condition set, same number of columns as the dleuth object design_matrix, minus the intercept).
replicate_dotplots(so,c(0,0))
# Note that many genes in the replicate samples have zero expression in one sample, but up to exp(4) in another.
# Exclude these low-expressed genes from a U test to reduce pval inflation caused by multiple testing correction (qval/FDR).
sleuth_u(so, "age", 65, target_ids=sleuth_reliable_target_ids(so, exp(4)))
}
