% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumshoe.R
\name{sleuth_u}
\alias{sleuth_u}
\title{Turn a numeric factor in a Sleuth object regression model into a binary one using a threshold value,
then perform the non-parametric U (a.k.a. Mann-Whitney or Wilcox) Test on the binarized factor.}
\usage{
sleuth_u(
  sleuth_obj,
  factor_name,
  value_threshold,
  target_ids = unique(sleuth_obj$obs_norm_filt$target_id)
)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep().}

\item{factor_name}{The existing numeric factor to be turned into 0 or 1 based on a threshold value.}

\item{value_threshold}{The threshold value for binarization (values for the given factor >= threshold are designated as "1", others are "0").}

\item{target_ids}{The list of transcripts the test should be run for (affects multiple testing correction as well).}
}
\value{
A table of \code{target_id}, \code{pval}, \code{qval} sorted by ascending pval.
}
\description{
For eukaryotic transcript analysis, tends to start generating multiple testing corrected p-values (FDR) <.05 when 10 or more samples
in each category (<= threshold, or > threshold) are present.
}
\examples{
# Read in metadata with "sample", "path" and other columns that can be used as regression model factors like "age."
meta <- read.table("tab_delimited_file_with sample_and_path_columns_etc.txt", header=TRUE)
meta$path <- as.character(meta$path)
so <- sleuth_prep(meta, ~treatment*age)
so <- sleuth_fit(so)
so <- sleuth_wt(so, "age")
wald_result <- sleuth_results(so, "age")
table(wald_result$qval<.05)
# The regression analysis shows few to no transcripts explained by age as a numeric linear dosage factor. What about <65 years old vs older?
wilcox_result <- sleuth_u(so, "age", 65)
table(wilcox_result$qval<.05)
}
\keyword{Mann-Whitney}
\keyword{U}
\keyword{Wilcox}
\keyword{non-parametric}
\keyword{test,}
\keyword{test.}
