% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_statistical_tests.R
\name{sleuth_wilcox_pairwise}
\alias{sleuth_wilcox_pairwise}
\title{Run a Wilcox pairwise comparison for a group across est_counts}
\usage{
sleuth_wilcox_pairwise(data, grouping, cutoff = 0.05)
}
\arguments{
\item{data}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit()}

\item{grouping}{A column name in the sample to covariate dataframe provided when running sleuth_prep / sleuth_interpret}

\item{cutoff}{The p-value cutoff to be used for the letter based output}
}
\value{
Results of a pairwise Wilcox test to identify the significantly different groups given as letters, where different letters indicate significance
}
\description{
Run a Wilcox pairwise comparison for a group across est_counts
}
\examples{
# Given a dataframe that contains est_counts for a particular gene and grouping information, run the Wilcox pairwise comparison of the est_counts factor between the selected group.
sleuth_kruskal_wallis(scaled_transcript_counts, "sex")
}
