% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{fdr_cutoff}
\alias{fdr_cutoff}
\title{Retrieve all target id's for a given data frame with a FDR less than the selected cutoff. Default cutoff value is .05.}
\usage{
fdr_cutoff(sleuth_res, q_cutoff = 0.05)
}
\arguments{
\item{sleuth_res}{A data frame generated following running sleuth_results}

\item{q_cutoff}{The selected FDR cutoff value. Default is .05.}
}
\value{
A list of target id's.
}
\description{
Retrieve all target id's for a given data frame with a FDR less than the selected cutoff. Default cutoff value is .05.
}
\examples{
# Given a Sleuth object, return a data frame of all target id's with an FDR less than the cutoff value
fdr_cutoff(wald_test_results)
}
