% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Generate a isoform level heat map. Can only show a maximum of three genes at the same time.}
\usage{
heatmap_plot(
  sleuth_obj,
  genes,
  q_max = 0.05,
  q_marker_vals = c(0.05, 0.01, 0.001),
  test = "wt",
  boxplot = TRUE
)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit()}

\item{genes}{A vector containing gene names whose transcripts should be plotted.}

\item{q_max}{The maximum q-value that should be used as a cutoff to determine significant transcripts}

\item{q_marker_vals}{A vector of the q-value cutoffs that should be used.}

\item{test}{The sleuth test to be analyzed using sleuth_results, can be either "wt" or "lrt". Defaults to "wt".}

\item{boxplot}{A Boolean value that indicates whether a box plot of scaled transcripts should be displayed.}
}
\description{
Generate a isoform level heat map. Can only show a maximum of three genes at the same time.
}
\examples{
# Create a isoform heat map from a sleuth_object. 
heatmap_plot(so, genes = c("Tp53", "Tnf", "Egfr", "Egr1"), q_max = .01, q_marker_vals = c(.05, .01, .001), test = "wt", bloxplot = FALSE)
}
