% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{self_plot_pca}
\alias{self_plot_pca}
\title{Generate a PCA plot from a sleuth object with user-selected colour_by and title.}
\usage{
self_plot_pca(
  obj,
  pc_x = 1L,
  pc_y = 2L,
  units = "est_counts",
  color_by = NULL,
  show_legend = TRUE,
  graph_name = "Sample Plot Name",
  auto_save = FALSE,
  scaled = TRUE
)
}
\arguments{
\item{obj}{A sleuth object.}

\item{pc_x}{Integer denoting the principle component to use for the x-axis.}

\item{pc_y}{Integer denoting the principle component to use for the y-axis.}

\item{units}{Either 'est_counts' ('scaled_reads_per_base' for gene_mode) or 'tpm'.}

\item{color_by}{Column name to colour the samples by. Default is NULL.}

\item{show_legend}{Boolean value if the legend should be shown in the figure out not.}

\item{graph_name}{Name of the graph. Default is "Sample Plot Name".}

\item{auto_save}{If TRUE, the plot will be saved automatically. Defaults to FALSE.}

\item{scaled}{Only works if auto_save is TRUE. The plot will be automatically saved and scaled to the amount of variation per axis. Scale factor is 1:5.}
}
\description{
Generate a PCA plot from a sleuth object with user-selected colour_by and title.
}
\examples{
# Create a PCA plot from a sleuth object.
self_plot_pca(so, colour_by = "tissue", graph_name = "All Samples")
}
