% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_lancaster_fc_merge}
\alias{sleuth_lancaster_fc_merge}
\title{Provides the merged results of two sleuth objects, one ran with gene aggregation and the second with p-value aggregation. Provides a fold-change (b) with the lancaster method.}
\usage{
sleuth_lancaster_fc_merge(sleuth_obj_1, sleuth_obj_2, q_max = NULL)
}
\arguments{
\item{sleuth_obj_1}{A sleuth object ran with either gene or p-value aggregation.}

\item{sleuth_obj_2}{A sleuth object ran with either gene or p-value aggregation.}

\item{q_max}{The FDR cutoff value that is passed to the fdr_cutoff function embedded within sleuth_object_result for sig_data. Defaults to NULL, which prevents it from running.}
}
\value{
A data frame containing the combined information from the two sleuth objects.
}
\description{
Provides the merged results of two sleuth objects, one ran with gene aggregation and the second with p-value aggregation. Provides a fold-change (b) with the lancaster method.
}
\examples{
sleuth_lancaster_fc_merge(so_gene, so_pVal, q_max = .05)
}
