% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_statistical_tests.R, R/gumshoe.R
\name{sleuth_model_rss}
\alias{sleuth_model_rss}
\title{Calculate and return the sum of the RSS values avalible for the model, model
factors, and the lrt factors (if the LRT was ran).}
\usage{
sleuth_model_rss(sleuth_obj)

sleuth_model_rss(sleuth_obj)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit().}
}
\value{
Dataframe with all the RSS values for a given sleuth_object.
}
\description{
Calculate and return the sum of the RSS values avalible for the model, model
factors, and the lrt factors (if the LRT was ran).

Calculate the Residual Sum of Squares of the Sleuth object's full regression model
}
\examples{
# Given a Sleuth object, retrieve the RSS values.
sleuth_model_rss(so_nominal)
}
