% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumshoe.R
\name{sleuth_model_rss}
\alias{sleuth_model_rss}
\title{Calculate the Residual Sum of Squares of the Sleuth object's full regression model}
\usage{
sleuth_model_rss(sleuth_obj)
}
\description{
Calculate the Residual Sum of Squares of the Sleuth object's full regression model
}
