% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_object_result}
\alias{sleuth_object_result}
\title{Retrieve all results from a given Sleuth object following statistical testing.}
\usage{
sleuth_object_result(
  sleuth_obj,
  all_data = FALSE,
  sig_data = TRUE,
  single_df = FALSE,
  retrived_from_model = FALSE,
  object_assignment = FALSE,
  q_max = 0.05,
  test = "wt"
)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit().}

\item{all_data}{A Boolean value that indicates whether all results of the test should be assigned to a variable. Automatic assignment to a variable in the global environment. Defaults to FALSE.}

\item{sig_data}{A Boolean value that indicates whether all the significant results of the test coupled with fdr_cutoff should be assigned to a variable. Automatic assignment to a variable in the global environment. Defaults to TRUE.}

\item{single_df}{A Boolean value that indicates whether all the results of the test should be merged into a single data frame and returned. Returned data must be manually assigned to variable when calling with single_df = TRUE. Defaults to FALSE.}

\item{retrived_from_model}{A Boolean value that indicates whether the data frame returned should include an additional column that includes the model that the data is from. This parameter is unique to single_df as both all_data and sig_data automatically assign data to variables named using model info in the global environment. Defaults to FALSE.}

\item{object_assignment}{A Boolean value that indicates whether the results retrieved when all_data or sig_data are TRUE should be assigned to the sleuth_object passed to the function.}

\item{q_max}{The FDR cutoff value that is passed to the fdr_cutoff function embedded within sleuth_object_result for sig_data. Defaults to .05.}

\item{test}{The sleuth test to be analyzed using sleuth_results, can be either "wt" or "lrt". Defaults to "wt".}
}
\value{
Either automatically generated variables (all_data and sig_data) containing resulting data from the Wald or Likewise Ratio tests with or without an FDR cutoff applied or all test result data in a manually assigned single data frame.
}
\description{
Retrieve all results from a given Sleuth object following statistical testing.
}
\examples{
# Given a Sleuth object and Boolean values of tests to run, perform the requested tests on all possible models for
# the given Sleuth object.
sleuth_object_result(so, q_max = .01, test = "lrt")
}
