% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_object_result}
\alias{sleuth_object_result}
\title{Retrieve all results from a given Sleuth object following statistical testing}
\usage{
sleuth_object_result(sleuth_obj, test = "wt")
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit()}

\item{test}{The sleuth test to be analyzed using sleuth_results, can be either "wt" or "lrt". Defaults to "wt".}
}
\value{
Firstly, newly generated variable containing resulting data from Wald or Likewise Ratio tests performed. Secondly, data frame containing data points with an FDR <.05
}
\description{
Retrieve all results from a given Sleuth object following statistical testing
}
\examples{
# Given a Sleuth object and Boolean values of tests to run, perform the requested tests on all possible models for the given Sleuth object
sleuth_object_result(so, test = "lrt")
}
