% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_sample_rss}
\alias{sleuth_sample_rss}
\title{Retrieve the RSS value for a selected model name and formula based on the analysis_data data frame.}
\usage{
sleuth_sample_rss(
  data,
  num_core = 1,
  selected_model = NULL,
  selected_formula = NULL
)
}
\arguments{
\item{data}{A data frame containing metadata file name, metadata tibble file, model names, and model formulae.}

\item{num_core}{An integer for the number of cores to be used for sleuth_prep.}

\item{selected_model}{A model name found in the analysis data frame, whose metadata will be used and samples removed from.}

\item{selected_formula}{The formula being used to run sleuth_prep on to assess the RSS.}
}
\value{
A data frame that contains a list of the samples removed and the associated RSS value.
}
\description{
Retrieve the RSS value for a selected model name and formula based on the analysis_data data frame.
}
\examples{
# Given a data frame containing a metadata_file_name, metadata_file, model_name, and model_data (formulae) alongside
# a specific model name and an associated selected_formula found in the data frame, run sleuth_prep() and sleuth_fit()
# and remove one sample in the metadata at a time and retrieve the sleuth object RSS value.
RSS_val <- sleuth_sample_rss(analysis_data, num_core = 3, selected_model = "dropped_nominal", selected_formula = ~sex*Treatment)
}
