% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{sleuth_test_wt}
\alias{sleuth_test_wt}
\title{Automated function to run all possible Wald tests on a given Sleuth object and the fitted model}
\usage{
sleuth_test_wt(sleuth_obj)
}
\arguments{
\item{sleuth_obj}{An existing Sleuth object as generated by sleuth_prep() and fit by sleuth_fit()}
}
\value{
Results of the Wald test assigned to the original Sleuth object input
}
\description{
Automated function to run all possible Wald tests on a given Sleuth object and the fitted model
}
\examples{
# Given a Sleuth object, run a Wald test on all possible models as determined on the formula used to run sleuth_prep()
sleuth_test_wt(so)
}
