% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleuth_pre_processing.R
\name{volc_plot}
\alias{volc_plot}
\title{Generate a volcano plot from a data frame containing a gene name, q-value, and fold change with user-selected cutoffs and title.}
\usage{
volc_plot(
  df,
  x_cutoff = 2,
  y_cutoff = -log10(0.05),
  graph_name = "Sample Graph Name"
)
}
\arguments{
\item{df}{A data frame or list containing ensemble transcript id's}

\item{x_cutoff}{X-axis cutoff, which corresponds to the fold change. Default value is 2.}

\item{y_cutoff}{Y-axis cutoff, which corresponds to the q-value. Default value is -log10(.05).}

\item{graph_name}{Name of the graph. Default is "Sample Graph Name".}
}
\description{
Generate a volcano plot from a data frame containing a gene name, q-value, and fold change with user-selected cutoffs and title.
}
\examples{
# Create a volcano plot from a data frame with gene symbols, q-value, and FC.
volc_plot(wald_sig_results, graph_name = "Treated v. WT - Sex Factor")
}
